requirements.txt
requests==2.31.0
python-dotenv==1.0.0
tiktoken==0.4.0
openai==0.27.8